make bindings work
