+++ packages/qwik-city/buildtime/vite/dev-server.ts
@@ -32,8 +32,12 @@ import {
 } from '../../middleware/request-handler/resolve-request-handlers';
 import { formatError } from './format-error';
 import { matchRoute } from 'packages/qwik-city/runtime/src/route-matcher';
+import { getWorkerdFunctions } from '../../../workerd-integration';
 
 export function ssrDevMiddleware(ctx: BuildContext, server: ViteDevServer) {
+  const fns = getWorkerdFunctions(server);
+  (globalThis as any).runLoader = fns.runLoader;
+
   const matchRouteRequest = (pathname: string) => {
     for (const route of ctx.routes) {
       let params = matchRoute(route.pathname, pathname);
@@ -76,6 +80,17 @@ export function ssrDevMiddleware(ctx: BuildContext, server: ViteDevServer) {
         }
       }
 
+      const pushModuleWithFilePath = (target: RouteModule[], filePath: string, module: RouteModule) => {
+        // routeModules needs to contain the modules as those are still needed in the
+        // rest of the middleware here, but such modules can't be serialized so we have
+        // no way of passing them to workerd, so let's just add an extra __filePath field
+        // here so that we re-initialize the module in workerd based on the filePath
+        target.push({
+          ...module,
+          __filePath: filePath,
+        } as RouteModule);
+      }
+
       const routeModulePaths = new WeakMap<RouteModule, string>();
       try {
         const { _deserializeData, _serializeData, _verifySerializable } =
@@ -87,6 +102,7 @@ export function ssrDevMiddleware(ctx: BuildContext, server: ViteDevServer) {
         const serverPlugins: RouteModule[] = [];
         for (const file of ctx.serverPlugins) {
           const layoutModule = await server.ssrLoadModule(file.filePath);
+          pushModuleWithFilePath(serverPlugins, file.filePath, layoutModule);
           serverPlugins.push(layoutModule);
           routeModulePaths.set(layoutModule, file.filePath);
           checkModule(loaderMap, layoutModule, file.filePath);
@@ -101,26 +117,15 @@ export function ssrDevMiddleware(ctx: BuildContext, server: ViteDevServer) {
           const route = routeResult.route;
           params = routeResult.params;
 
-          const pushModuleWithFilePath = (filePath: string, module: RouteModule) => {
-            // routeModules needs to contain the modules as those are still needed in the
-            // rest of the middleware here, but such modules can't be serialized so we have
-            // no way of passing them to workerd, so let's just add an extra __filePath field
-            // here so that we re-initialize the module in workerd based on the filePath
-            routeModules.push({
-              ...module,
-              __filePath: filePath,
-            } as RouteModule);
-          }
-
           // found a matching route
           for (const layout of route.layouts) {
             const layoutModule = await server.ssrLoadModule(layout.filePath);
-            pushModuleWithFilePath(layout.filePath, layoutModule);
+            pushModuleWithFilePath(routeModules, layout.filePath, layoutModule);
             routeModulePaths.set(layoutModule, layout.filePath);
             checkModule(loaderMap, layoutModule, layout.filePath);
           }
           const endpointModule = await server.ssrLoadModule(route.filePath);
-          pushModuleWithFilePath(route.filePath, endpointModule);
+          pushModuleWithFilePath(routeModules, route.filePath, endpointModule);
           routeModulePaths.set(endpointModule, route.filePath);
           checkModule(loaderMap, endpointModule, route.filePath);
         }
