+++ packages/qwik-city/buildtime/vite/workerd-integration.ts
@@ -1,0 +1,1 @@
+export { getWorkerdFunctions } from '../../../workerd-integration';
