+++ packages/qwik-city/package.json
@@ -41,7 +41,9 @@
     "unified": "10.1.2",
     "unist-util-visit": "5.0.0",
     "uvu": "0.5.6",
-    "yaml": "^2.3.2"
+    "yaml": "^2.3.2",
+    "miniflare": "https://prerelease-registry.devprod.cloudflare.dev/workers-sdk/runs/6732252267/npm-package-miniflare-4322",
+    "workerd-vite-utils": "/Users/dario/Repos/my-repos/preact-vite-workerd-ssr/workerd-vite-utils"
   },
   "engines": {
     "node": ">=16.8.0 <18.0.0 || >=18.11"
