+++ scripts/qwik-city.ts
@@ -234,6 +234,7 @@ async function buildVite(config: BuildConfig) {
     'typescript',
     'vite-imagetools',
     'svgo',
+    'miniflare'
   ];
 
   const swRegisterPath = join(config.srcQwikCityDir, 'runtime', 'src', 'sw-register.ts');
